requirements.txt
flask==2.2.2
flask-sqlalchemy==3.0.2
psycopg2-binary
requests 